+++ NEWS.md
@@ -20,6 +20,9 @@
 
 * `tab_footnote()` + `md()` + docx is now processed correctly (@olivroy, #1892).
 
+* Added support for some unicode characters when going out to LaTeX (@thebioengineer)
+
+
 # gt 1.0.0
 
 ## Minor improvements and bug fixes
@@ -41,7 +44,6 @@
 * `gt(row_group_as_column = TRUE)` now works if multiple groups are supplied (#1552). (@olivroy, #1801).
 
 * Fixed many typos in the documentation. (#1910, thanks @MichaelChirico!)
-
 
 # gt 0.11.1